[#157113826] refactor app
