+++ api/tests/test_asset_api.py
@@ -4,7 +4,7 @@ from django.contrib.auth import get_user_model
 from rest_framework.reverse import reverse
 from rest_framework.test import APIClient
 
-from ..models import Asset, AssetModelNumber, SecurityUser, AssetLog
+from core.models import Asset, AssetModelNumber, SecurityUser, AssetLog
 
 User = get_user_model()
 client = APIClient()
