+++ core/models/__init__.py
@@ -1,0 +1,2 @@
+from .user import *  # noqa: F403,F401
+from .asset import *  # noqa: F403,F401
