+++ core/tests/test_allocation_history_model.py
