+++ core/tests/test_asset_category_model.py
