+++ core/tests/test_asset_model.py
