+++ core/tests/test_asset_model_number.py
