+++ core/tests/test_asset_status_model.py
