+++ core/tests/test_asset_type_model.py
